apache-airflow/requirements.txt
# base
apache-airflow==2.0.1
flake8==3.9.0
pre-commit==2.11.1

# providers
apache-airflow-providers-amazon==1.2.0
apache-airflow-providers-postgres==1.0.1
apache-airflow-providers-cncf-kubernetes==1.0.2

# extra
minio==7.0.2
pandas==1.2.3
SQLAlchemy==1.4.2
pyspark==3.0.0